Move CommitLogParams to params package
