+++ servers/quarkus-server/src/test/java/org/projectnessie/server/TestRest.java
@@ -48,8 +48,8 @@ import org.junit.jupiter.api.function.Executable;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.projectnessie.api.CommitLogParams;
-import org.projectnessie.api.CommitLogParams.Builder;
+import org.projectnessie.api.params.CommitLogParams;
+import org.projectnessie.api.params.CommitLogParams.Builder;
 import org.projectnessie.api.ContentsApi;
 import org.projectnessie.api.TreeApi;
 import org.projectnessie.client.NessieClient;
