+++ servers/services/src/main/java/org/projectnessie/services/rest/TreeResource.java
@@ -27,7 +27,7 @@ import java.util.stream.Stream;
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 
-import org.projectnessie.api.CommitLogParams;
+import org.projectnessie.api.params.CommitLogParams;
 import org.projectnessie.api.TreeApi;
 import org.projectnessie.error.NessieConflictException;
 import org.projectnessie.error.NessieNotFoundException;
